requirements.txt
# CV RAG System - Requirements
# =============================

# Core dependencies
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.4.0
openai>=1.0.0

# Data processing
pandas>=2.0.0
pyarrow>=14.0.0

# PDF extraction
pypdf>=4.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Web UI
streamlit>=1.30.0

# Visualization
matplotlib>=3.7.0

# Optional: Local embeddings
# sentence-transformers>=2.2.0

# Utilities
tqdm>=4.65.0
